Fix promote_int mode 3 to match LP64 layout
Mode 3 set mw_use_int64_t for long and mw_use_uint32_t for uint but
returned int32_t and uint64_t respectively, so the emitted typename
contradicted the typedef forward-declarations.  Under LP64 (where
int/uint stay 32-bit and long/ulong become 64-bit), the intended
mapping is long -> int64_t and uint -> uint32_t.  Also noted the
layout in a short comment for future readers.

+++ python/mwrap_ast.py
@@ -187,13 +187,14 @@ def promote_int(ctx, name: str) -> str:
         if name == "uint":  return "ulong"
         if name == "ulong": return "ulong"
     elif ctx.mw_promote_int == 3:
+        # LP64: int/uint stay 32-bit, long/ulong become 64-bit.
         if name == "int":   ctx.mw_use_int32_t = 1
         if name == "long":  ctx.mw_use_int64_t = 1
         if name == "uint":  ctx.mw_use_uint32_t = 1
         if name == "ulong": ctx.mw_use_uint64_t = 1
         if name == "int":   return "int32_t"
-        if name == "long":  return "int32_t"
-        if name == "uint":  return "uint64_t"
+        if name == "long":  return "int64_t"
+        if name == "uint":  return "uint32_t"
         if name == "ulong": return "uint64_t"
     elif ctx.mw_promote_int == 4:
         if name == "int":   ctx.mw_use_int64_t = 1
